Prevent password hashes from ever touching emergency
