+++ back/core/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "core"
-version = "0.1.76"
+version = "0.1.77"
 description = ""
 authors = [
     {name = "Your Name",email = "you@example.com"}
