+++ back/core/src/core/domain/entities/historical_emergency.py
@@ -5,26 +5,9 @@ from typing import Dict
 
 from core.domain.entities.emergency import Emergency, EmergencyStatus
 from core.domain.entities.medical_center import MedicalCenterInfo
-from core.domain.entities.user import User, UserRole
+from core.domain.entities.user import GeneralUser, UserRole
 from core.domain.value_objects.location import Location
 from core.domain.value_objects.triage import Triage
-
-
-@dataclass
-class HistoricalUser:
-    id: uuid.UUID
-    name: str
-    email: str
-    userRole: UserRole
-
-    @classmethod
-    def from_user(cls, user: User):
-        return cls(
-            id=user.id,
-            name=user.name,
-            email=user.email,
-            userRole=user.userRole,
-        )
 
 
 @dataclass
@@ -37,8 +20,8 @@ class HistoricalEmergency:
     filingNumber: int
     triage: Triage | None
     finalStatus: EmergencyStatus
-    operatedBy: HistoricalUser | None
-    assignedTo: HistoricalUser | None
+    operatedBy: GeneralUser | None
+    assignedTo: GeneralUser | None
     transferedTo: MedicalCenterInfo | None
     cancelReason: str | None
     timeline: Dict[EmergencyStatus, datetime]
@@ -52,12 +35,8 @@ class HistoricalEmergency:
             triage=emergency.triage,
             finalStatus=emergency.status,
             transferedTo=emergency.transferedTo,
-            operatedBy=HistoricalUser.from_user(emergency.operatedBy)
-            if emergency.operatedBy is not None
-            else None,
-            assignedTo=HistoricalUser.from_user(emergency.assignedTo)
-            if emergency.assignedTo is not None
-            else None,
+            operatedBy=emergency.operatedBy,
+            assignedTo=emergency.assignedTo,
             cancelReason=emergency.cancelReason,
             timeline=emergency.timeline,
         )
